chore: move message.rs to core.rs
