+++ core/src/services/huggingface/lister.rs
@@ -20,8 +20,8 @@ use std::sync::Arc;
 use async_trait::async_trait;
 
 use super::core::HuggingfaceCore;
+use super::core::HuggingfaceStatus;
 use super::error::parse_error;
-use super::message::HuggingfaceStatus;
 use crate::raw::*;
 use crate::*;
